WIP Improve README's clarity of example data flow
I'm not sure if this is really better or just wordier, but when glancing it over I felt like it just starts out using TodoList and TodoStore without really showing how they fit together in relation to RiotControl itself and how simple they are. I think most people will get it, esp. if they are familiar with React but new people might not.